Kraken_api_trading_bot
Kraken is a cryptocurrency exchange based in San Francisco where market participants can trade various cryptocurrencies bought or sold with fiat currencies. As of June 2024, Kraken was ranked sixth by CoinMarketCap for spot trading and fifth for derivatives trading. It had an average daily trading volume of roughly $281 million.This code help you to connect with Kraken api and implement your strategies as well as backtest it